requirements.txt
numpy
wget
bounter
internetarchive
beautifulsoup4>=4.10.0
lxml
xopen
nbimporter
nbformat
pandas
ujson
xlrd
xlwt
htrc-feature-reader
geopy
moviepy
statsmodels
plotnine
networkx
ujson
yapmap
loguru
sacremoses
sklearn
nltk>=3.6.4
smart_open
tabulate
mgzip
openpyxl
ftfy
dateparser
python-dotenv
blosc
cryptography
ordered_set
orjson
tinydb
cassandra-driver
recordlinkage
datasketch
sqlitedict
rispy
wptools
unidecode